+++ canvas/canvas-graph.js
@@ -1,0 +1,4 @@
+import { CanvasGraph } from './CanvasGraph/CanvasGraph.js';
+
+export { CanvasGraph };
+export default CanvasGraph;